Added ttsc and abstracted imports to absolute paths
