+++ package.json
@@ -6,7 +6,7 @@
   "bin": "./dist/server/app.js",
   "scripts": {
     "start": "node ./dist/server/app.js",
-    "build": "npx rimraf ./dist/* && (ng build & tsc --p ./tsconfig.server.json)",
+    "build": "npx rimraf ./dist/* && (ng build & ttsc --p ./tsconfig.server.json)",
     "package": "npx rimraf ./release/* && npx pkg . --out-path ./release && create-nodew-exe ./release/tvs-cp.exe ./release/tvs-cp-headless.exe",
     "postinstall": "npm explore memoryjs -- npm run build64",
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -47,8 +47,8 @@
     "@angular/cli": "~9.0.2",
     "@angular/compiler-cli": "~9.0.1",
     "@angular/language-service": "~9.0.1",
-    "@types/node": "^12.12.28",
     "@types/fast-levenshtein": "0.0.1",
+    "@types/node": "^12.12.28",
     "@types/ws": "^7.2.1",
     "awesome-typescript-loader": "^5.2.1",
     "codelyzer": "^5.1.2",
@@ -59,7 +59,9 @@
     "rimraf": "^3.0.2",
     "ts-node": "~8.3.0",
     "tslint": "^6.0.0",
-    "typescript": "~3.7.5"
+    "ttypescript": "^1.5.10",
+    "typescript": "~3.7.5",
+    "typescript-transform-paths": "^1.1.14"
   },
   "pkg": {
     "assets": [
