+++ src/common/states/EmulatorState.ts
@@ -1,0 +1,7 @@
+export enum EmulatorState {
+  NOT_CONNECTED = 'NOT_CONNECTED',
+  CONNECTING = 'CONNECTING',
+  CONNECTED = 'CONNECTED',
+  PATCHING = 'PATCHING',
+  PATCHED = 'PATCHED',
+}
