+++ tsconfig.server.json
@@ -2,7 +2,7 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "baseUrl": "./",
-    "outDir": "./dist/server",
+    "outDir": "./dist",
     "target": "es2016",
     "module": "commonjs",
     "moduleResolution": "node",
